requirements.txt
mcp
fastmcp
psycopg2-binary
python-dotenv